+++ App/src/components/MemberSectionList.tsx
@@ -56,6 +56,7 @@ enum ItemType {
 }
 
 const logger = new Logger("MemberSectionList");
+const SCREEN_FACTOR = Dimensions.get("screen").width / 375 /* 6S */;
 
 export class MemberSectionListBase extends React.Component<Props, State>  {
     dataProvider: DataProvider;
@@ -87,8 +88,8 @@ export class MemberSectionListBase extends React.Component<Props, State>  {
                     // this kills the performance of the "debugger" as
                     // code is executed many many times in every render cycle
                     // if (__DEV__) { logger.debug(roles, length); }
-                    if (length > 80) return ItemType.Large3;
-                    if (length > 43) return ItemType.Large2;
+                    if (length > 80 * SCREEN_FACTOR) return ItemType.Large3;
+                    if (length > 47 * SCREEN_FACTOR) return ItemType.Large2;
 
                     return ItemType.Large;
                 }