+++ Backend/src/graphql/cache/TTLs.ts
@@ -2,12 +2,13 @@ const HOUR = 60*60;
 const hours = (hours: number) => hours * HOUR;
 
 export const TTLs = {
-    MemberOverview: hours(8),   // base details don't change often
-    ClubMembers:    hours(8),   // base details don't change often
+    MemberOverview: hours(8),       // base details don't change often
+    ClubMembers:    hours(8),       // base details don't change often
 
-    Member: 0,                  // updated by cache
-    Structure: 0,               // removed by cache maintenance
+    Member: 0,                      // updated by cache
+    Structure: 0,                   // removed by cache maintenance
 
+    StructureOverview: hours(24),
     Albums: hours(4),
 
     // everything else e.g.