+++ Backend/src/schedule/writeValues.ts
@@ -2,13 +2,24 @@ import { Client } from 'pg';
 import { removeEmpty } from './removeEmpty';
 import { Types } from './types';
 
+const clubPK = (c: any) => c["subdomain"].replace(/[^a-z]/ig, "") + "_" + c["subdomain"].replace(/[^0-9]/ig, "");
+const memberPK = (c: any) => c["id"];
+
+export type ChangePointer = {
+    id: string | number,
+    type: Types,
+};
+
 export const writeValues = (client: Client, type: Types) => {
-    return async (data: Array<any>) => {
-        console.log("Writing chunk of", data.length, "records");
+    const pk = type === Types.clubs ? clubPK : memberPK;
+
+    return async (data: Array<any>): Promise<ChangePointer[]> => {
+        console.log("Writing chunk of", data.length, type, "records");
+
         const results: any[] = [];
 
         for (let r of data) {
-            const id = r.id || r.subdomain;
+            const id = pk(r);
             console.log(id);
 
             const result = await client.query(`
@@ -22,7 +33,7 @@ DO UPDATE
 
             if (result.rowCount == 1) {
                 console.log(id, "modified");
-                results.push(id);
+                results.push({ id, type });
             }
         }