+++ App/assets/docs/releasenotes.md
@@ -3,6 +3,8 @@ Release candidate.
 
 ## Improvements
 - Shake to report a problem
+- Filters for association board and -assists
+- Move login buttons with when KeyBoard is visible
 
 ## Fixes
 - Toggeling favorites does not immediately update members view
@@ -10,6 +12,7 @@ Release candidate.
 - Hide roles section on profile page if there are no entries in the list
 - Updated external licenses MD
 - Added about page to legal
+- Added screenfactor for iPad devices that roles break correctly
 
 ## Operations
 - Moved from writethrough to active cache updates